Initialize react-query client and refactor GET request to use react-query
